pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alfred-vault"
version = "0.3.2"
description = "Personal agentic infrastructure. Self-hosted. Always on."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-frontmatter>=1.1",
    "structlog>=24.0",
    "watchdog>=4.0",
    "rich>=13.0",
    "textual>=0.89",
]

[project.optional-dependencies]
temporal = ["temporalio>=1.9.0"]
surveyor = [
    "pymilvus[milvus_lite]>=2.4",
    "scikit-learn>=1.4",
    "leidenalg>=0.10",
    "igraph>=0.11",
    "numpy>=1.26",
    "openai>=1.30",
]
voice = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]
all = ["alfred-vault[surveyor]", "alfred-vault[temporal]", "alfred-vault[voice]"]

[project.urls]
Homepage = "https://github.com/ssdavidai/alfred"
Repository = "https://github.com/ssdavidai/alfred"

[project.scripts]
alfred = "alfred.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/alfred"]

[tool.hatch.build.targets.sdist]
include = ["src/alfred/", "README.md", "LICENSE"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra"
